Criando metodosAuxiliares.js para escolas/Despesas
